+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+- Read from standard input in the CLI ([#48](https://github.com/gadomski/antimeridian/pull/48))
+
 ## [0.3.0] - 2023-06-29
 
 ### Changed